some QSQL implementations

+++ Scene/include/Call.h
@@ -11,10 +11,10 @@
 namespace moe {
 
     struct Call{
-        int id_, eid_, symbol_address_, start_time_, total_time_;
+        uint64_t id_, eid_, symbol_address_, start_time_, total_time_;
         std::string symbol_name_;
         QVector<Call*> children;
-        Call(int id, int eid, int symbol_address, int start_time, int total_time , std::string symbol_name) :
+        Call(uint64_t id, uint64_t eid, uint64_t symbol_address, uint64_t start_time, uint64_t total_time , std::string symbol_name) :
                 id_(id), eid_(eid), symbol_address_(symbol_address), start_time_(start_time), total_time_(total_time),
                 symbol_name_(symbol_name) {}
     };
